Task:0 0-putchar.c :white_check_mark:

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -1,18 +1,21 @@
 #include "main.h"
-/*
+/**
+ * main - Entry point
+ *
+ * Description: it outputs '_putchar'
+ *
+ * Return: Always (0)
  */
 int main(void)
 {
-        char word[8] = "_putchar";
-        int i = 0;
+	char word[8] = "_putchar";
+	int i = 0;
 
-        while(i < sizeof(word))
-        {
-                _putchar(word[i]);
-                i++;
-        }
-        _putchar('\n');
-
-        return (0);
-
+	while (i < (int)sizeof(word))
+	{
+		_putchar(word[i]);
+		i++;
+	}
+	_putchar('\n');
+	return (0);
 }
